requirements.txt
google-cloud-logging==1.12.1
google-cloud-storage==1.19.1
python-json-logger==0.1.11
pyppeteer==0.0.25
requests==2.21.0
